more stuff

+++ resources/shaders/hlsl/vulkan_exp/clear_tex_to_color.comp.hlsl
@@ -9,17 +9,17 @@ struct PC {
 
 PUSHCONSTANT(PC, pc);
 
-RWTexture2D<float4> out_tex : register(u0);
+RWTexture2D<min16float4> out_tex : register(u0);
 
 [RootSignature(ROOT_SIGNATURE)][NumThreads(8, 8, 1)] void main(uint2 dtid : SV_DispatchThreadID) {
   if (dtid.x >= pc.dims.x || dtid.y >= pc.dims.y) return;
 
   if (dtid.x < pc.dims.x / 2) {
     if (dtid.y < pc.dims.y / 2) {
-      out_tex[dtid] = float4(1.0, 0.0, 0.0, 1.0);
+      out_tex[dtid] = min16float4(1.0, 0.0, 0.0, 1.0);
 
     } else {
-      out_tex[dtid] = float4(0.0, 1.0, 1.0, 1.0);
+      out_tex[dtid] = min16float4(0.0, 1.0, 1.0, 1.0);
     }
   }
 }
